+++ src/v3/app/App.js
@@ -1,31 +1,32 @@
 // @flow
 
 import React from "react";
-import {Route, NavLink, type Match} from "react-router-dom";
+import {BrowserRouter as Router, Route, NavLink} from "react-router-dom";
 
 import CredExplorer from "./credExplorer/App";
 
-export default class App extends React.Component<{match: Match}> {
+export default class App extends React.Component<{}> {
   render() {
-    const {match} = this.props;
-    const CRED_EXPLORER_ROUTE = match.url + "/explorer";
+    const CRED_EXPLORER_ROUTE = "/explorer";
     return (
-      <div>
-        <nav>
-          <ul>
-            <li>
-              <NavLink to={match.url}>Home</NavLink>
-            </li>
-            <li>
-              <NavLink to={CRED_EXPLORER_ROUTE}>Cred Explorer</NavLink>
-            </li>
-          </ul>
-        </nav>
+      <Router>
+        <React.Fragment>
+          <nav>
+            <ul>
+              <li>
+                <NavLink to="/">Home</NavLink>
+              </li>
+              <li>
+                <NavLink to={CRED_EXPLORER_ROUTE}>Cred Explorer</NavLink>
+              </li>
+            </ul>
+          </nav>
 
-        <hr />
-        <Route exact path={match.url} component={Home} />
-        <Route path={CRED_EXPLORER_ROUTE} component={CredExplorer} />
-      </div>
+          <hr />
+          <Route exact path="/" component={Home} />
+          <Route path={CRED_EXPLORER_ROUTE} component={CredExplorer} />
+        </React.Fragment>
+      </Router>
     );
   }
 }
