+++ src/v3/app/index.js
@@ -1,0 +1,13 @@
+// @flow
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import App from "./App";
+import registerServiceWorker from "./registerServiceWorker";
+
+const root = document.getElementById("root");
+if (root == null) {
+  throw new Error("Unable to find root element!");
+}
+ReactDOM.render(<App />, root);
+registerServiceWorker();
