+++ src/v3/app/registerServiceWorker.js
